MyProject.Test/API.Test/Controller.Test/PermissionGroupControllerTests.cs
using System;
using System.Collections.Generic;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Mvc;
using Moq;
using MyProject.Api.Controllers;
using MyProject.Application.Interfaces;
using MyProject.Domain.Entities;
using Xunit;

namespace MyProject.Tests.Controllers
{
    public class PermissionGroupControllerTests
    {
        [Fact]
        public async Task GetAllPermissionGroups_ReturnsOkResult()
        {
            // Arrange
            var mockService = new Mock<IPermissionGroupService>();
            mockService.Setup(service => service.GetAllAsync()).ReturnsAsync(new List<PermissionGroup>());

            var controller = new PermissionGroupController(mockService.Object);

            // Act
            var result = await controller.GetAllPermissionGroups();

            // Assert
            Assert.IsType<OkObjectResult>(result);
        }

        [Theory]
        [InlineData(1)]
        [InlineData(2)]
        public async Task GetPermissionGroupById_ExistingId_ReturnsOkResult(int permissionGroupId)
        {
            // Arrange
            var mockService = new Mock<IPermissionGroupService>();
            mockService.Setup(service => service.GetByIdAsync(permissionGroupId)).ReturnsAsync(new PermissionGroup { Id = permissionGroupId, Name = "TestPermissionGroup" });

            var controller = new PermissionGroupController(mockService.Object);

            // Act
            var result = await controller.GetPermissionGroupById(permissionGroupId);

            // Assert
            Assert.IsType<OkObjectResult>(result);
        }

        [Fact]
        public async Task CreatePermissionGroup_ValidObject_ReturnsCreatedAtAction()
        {
            // Arrange
            var mockService = new Mock<IPermissionGroupService>();
            var newPermissionGroup = new PermissionGroup { Name = "NewPermissionGroup" }; // Replace with test data
            var createdPermissionGroupId = 1; // Replace with an expected ID for testing purposes
            mockService.Setup(service => service.CreateAsync(newPermissionGroup)).ReturnsAsync(createdPermissionGroupId);

            var controller = new PermissionGroupController(mockService.Object);

            // Act
            var result = await controller.CreatePermissionGroup(newPermissionGroup) as CreatedAtActionResult;

            // Assert
            Assert.NotNull(result);
            Assert.Equal(nameof(PermissionGroupController.GetPermissionGroupById), result.ActionName);
            Assert.Equal(createdPermissionGroupId, result.RouteValues["id"]);
        }

        // Write similar tests for UpdatePermissionGroup and DeletePermissionGroup actions

        // Remember to add specific assertions to validate the responses and behaviors of each action.
    }
}
